feat: [ANDROAPP-5125] remove old code
