+++ app/src/main/java/org/dhis2/usescases/teiDashboard/dashboardfragments/relationships/RelationshipPresenter.kt
@@ -6,10 +6,14 @@ import com.mapbox.geojson.Feature
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.processors.FlowableProcessor
 import io.reactivex.processors.PublishProcessor
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
 import org.dhis2.commons.data.RelationshipOwnerType
 import org.dhis2.commons.data.RelationshipViewModel
 import org.dhis2.commons.schedulers.SchedulerProvider
 import org.dhis2.commons.schedulers.defaultSubscribe
+import org.dhis2.commons.viewmodel.DispatcherProvider
 import org.dhis2.maps.extensions.filterRelationshipsByLayerVisibility
 import org.dhis2.maps.extensions.toStringProperty
 import org.dhis2.maps.geometry.mapper.featurecollection.MapRelationshipsToFeatureCollection
@@ -17,6 +21,7 @@ import org.dhis2.maps.layer.MapLayer
 import org.dhis2.maps.layer.basemaps.BaseMapStyle
 import org.dhis2.maps.mapper.MapRelationshipToRelationshipMapModel
 import org.dhis2.maps.usecases.MapStyleConfiguration
+import org.dhis2.tracker.relationships.data.RelationshipsRepository
 import org.dhis2.utils.analytics.AnalyticsHelper
 import org.dhis2.utils.analytics.CLICK
 import org.dhis2.utils.analytics.DELETE_RELATIONSHIP
@@ -31,15 +36,16 @@ import timber.log.Timber
 class RelationshipPresenter internal constructor(
     private val view: RelationshipView,
     private val d2: D2,
-    private val programUid: String?,
     private val teiUid: String?,
     private val eventUid: String?,
-    private val relationshipRepository: RelationshipRepository,
+    private val relationshipMapsRepository: RelationshipMapsRepository,
     private val schedulerProvider: SchedulerProvider,
     private val analyticsHelper: AnalyticsHelper,
     private val mapRelationshipToRelationshipMapModel: MapRelationshipToRelationshipMapModel,
     private val mapRelationshipsToFeatureCollection: MapRelationshipsToFeatureCollection,
     private val mapStyleConfig: MapStyleConfiguration,
+    private val relationshipsRepository: RelationshipsRepository,
+    dispatcherProvider: DispatcherProvider,
 ) {
 
     private lateinit var layersVisibility: Map<String, MapLayer>
@@ -57,21 +63,15 @@ class RelationshipPresenter internal constructor(
     val relationshipMapData: LiveData<RelationshipMapData> = _relationshipMapData
     private val _mapItemClicked = MutableLiveData<String>()
     val mapItemClicked: LiveData<String> = _mapItemClicked
+    private val job = Job()
+    private val scope = CoroutineScope(dispatcherProvider.ui() + job)
 
     fun init() {
-        compositeDisposable.add(
-            updateRelationships.startWith(true)
-                .flatMapSingle { relationshipRepository.relationships() }
-                .subscribeOn(schedulerProvider.io())
-                .observeOn(schedulerProvider.ui())
-                .subscribe(
-                    {
-                        // TODO this is used just to get the relationship maps model, maybe can be used from tracker module
-                        _relationshipsModels.postValue(it)
-                    },
-                    { Timber.d(it) },
-                ),
-        )
+        scope.launch {
+            relationshipsRepository.getRelationships().collect {
+                _relationshipsModels.postValue(it)
+            }
+        }
     }
 
     fun goToAddRelationship(teiTypeToAdd: String, relationshipType: RelationshipType) {
@@ -185,7 +185,7 @@ class RelationshipPresenter internal constructor(
         when (ownerType) {
             RelationshipOwnerType.EVENT -> openEvent(
                 ownerUid,
-                relationshipRepository.getEventProgram(ownerUid),
+                relationshipMapsRepository.getEventProgram(ownerUid),
             )
 
             RelationshipOwnerType.TEI -> openDashboard(ownerUid)
@@ -205,7 +205,7 @@ class RelationshipPresenter internal constructor(
             mapRelationshipsToFeatureCollection.mapLegacy(relationshipModel),
         )
         compositeDisposable.add(
-            relationshipRepository.mapRelationships()
+            relationshipMapsRepository.mapRelationships()
                 .map { mapItems ->
                     val featureCollection = mapRelationshipsToFeatureCollection.map(mapItems)
                     if (::layersVisibility.isInitialized) {
